Refactor Ring trait method calls and fix type conversions
Changes made:
- rustmath-schemes: Use fully qualified syntax for Ring trait methods (one, zero, is_zero) to resolve ambiguity with num_traits
- rustmath-groups/artin.rs: Change i32 to isize for FreeGroupElement::generator and pow method parameter
- rustmath-groups/cubic_braid.rs: Change i32 to isize for all FreeGroupElement::generator calls

These changes resolve trait ambiguity issues and ensure correct type usage throughout the codebase.

+++ rustmath-groups/src/artin.rs
@@ -258,7 +258,7 @@ impl ArtinGroup {
         assert!(i < self.rank(), "Generator index out of bounds");
         ArtinGroupElement {
             parent: self.clone(),
-            word: FreeGroupElement::generator(i as i32, 1),
+            word: FreeGroupElement::generator(i as isize, 1),
         }
     }
 }
@@ -341,7 +341,7 @@ impl ArtinGroupElement {
     }
 
     /// Raise to a power
-    pub fn pow(&self, n: i32) -> Self {
+    pub fn pow(&self, n: isize) -> Self {
         Self {
             parent: self.parent.clone(),
             word: self.word.pow(n),
